requirements.txt
deep-translator==1.11.4
python-telegram-bot==21.4
